[refact] Move JSON models inside network module
