+++ src/client/property_snapshot.rs
@@ -83,7 +83,7 @@ impl Property for PropertySnapshot {
 pub mod tests {
 
     use super::*;
-    use crate::models::{ConfigValue, Rule, Segment, SegmentRule, Segments, ValueType};
+    use crate::network::models::{ConfigValue, Rule, Segment, SegmentRule, Segments, ValueType};
     use std::collections::HashMap;
 
     #[test]
