+++ src/network/mod.rs
@@ -25,3 +25,4 @@ pub(crate) mod live_configuration;
 
 pub use errors::NetworkError;
 pub type NetworkResult<T> = std::result::Result<T, NetworkError>;
+pub(crate) mod models;
