+++ src/tests/test_get_feature.rs
@@ -14,7 +14,7 @@
 
 use std::collections::HashMap;
 
-use crate::models::ConfigurationJson;
+use crate::network::models::ConfigurationJson;
 
 use crate::client::configuration::Configuration;
 use crate::{AppConfigurationClient, ConfigurationProvider};
@@ -23,7 +23,7 @@ use rstest::*;
 
 use super::client_enterprise;
 use crate::feature::Feature;
-use crate::models::tests::configuration_unordered_segment_rules;
+use crate::network::models::tests::configuration_unordered_segment_rules;
 
 #[rstest]
 fn test_get_feature_doesnt_exist(client_enterprise: Box<dyn AppConfigurationClient>) {
